Add resize! method to typed arrays.
This does something that I'm not 100% sure about by modifying the data
member of an RData, but I don't see any reason to believe it's not a
reasonable course of action if the underlying pointer changes, provided
the object is fundamentally the same.

At any rate, this is just a simple realloc (well, REALLOC_N, which isn't
specifically realloc, but you know what I mean). New elements of an
array may be garbage.