Swap use of int for size_t when referring to indices or lengths.
This just seems like a Good Thing to do since it's pretty much the best
use of size_t. Leaving in the checks for less than zero for now, though,
as it'll be optimized out and if I need to change it back later, the
bounds checking is still there.