Account for loss of cache when calling ___Array#resize!
I might also add some sort of call to cached objects to allow them to
update their pointers, but for now, this will do.

I should probably move the caching code and such to C, but being able to
handle it in Ruby is fairly convenient. I'd prefer to only drop down to
C for things that would be called fairly often, and resizing a typed
array shouldn't be one of those things.