Add optimization flags.
fast-math and strict-aliasing are both disabled despite use of -O3.
Although fast-math isn't specified by -O3, I do want to make sure it's
completely disabled so that the operations are IEEE-compliant, or as
compliant as I can hope for. This comes at the cost of some performance,
but it results in somewhat-more-deterministic floating point values, so
I think it's a good trade-off. I'm not sure if this will actually have
much effect on Windows, but then this code isn't tested on Windows and
that makes it hard to say what'll happen.

strict-aliasing is disabled because it just plain breaks code in weird,
hard to debug ways, and I don't want to go through that when it's not
entirely necessary.