Make extconf.rb executable.
This is mainly for testing, though can be handy if you for some reason
want to go in and rebuild something yourself. I don't anticipate anyone
doing that, but there are crazy people out there.

In this case, I just happen to need it for building the bindings so I
can check for compile errors before building/installing the gem.