Fix copy/paste issue in sm_mat3_init.
Convenient as it is, I need to remember to fix these things.

Prior to this, mat3_init was the same as mat4_init, except with a mat3_t
- so that led to bad memory access, where if you provided an array it
was possible to write outside the bounds of the mat3. In addition, if
you provided all components of the mat3 (all 16, because it was for
mat4) you'd also do evil things. Same for the Vec4 init.

So, now it only takes three Vec3s, reads only at most 9 elements from an
array, and otherwise takes 9 component arguments.