Move back to support modules for inspect.
I don't particularly like doing this because it means a bunch of lines
that essentially do the same thing multiple times for differently-named
classes, but it's the only easy way to provide documentation. Turns out
rdoc/yard are both very, very bad at handling documentation if it's not
inline with the source code. So, back to the slightly less convenient
modules for the sake of generating documentation.