Kill off a lot of macro usage and update the bindings a bit.
So, notable changes:

- Documentation! Because the preprocessor bits are mostly gone. At least
  there's docs in the C bindings. The Ruby side of things still needs to
  be documented.

- Array types no longer depends on array_cache.rb to cache their fetched
  data wrappers and checks for the existence and so on in the C
  bindings. This might be faster than the Ruby bindings, but the main
  up-shot is that this removes a hack that I layered on top of the typed
  arrays.

- dup and clone are now just aliases of copy. The default dup/clone
  behaviour isn't beneficial for any of the math types and could lead
  very severe problems, so I consider this necessary. This means that
  initalize_copy will never be called by any math type subclass.

These are all more or less motivated by the desire to get some manner of
inline documentation for the source without resorting to really weird
hacks. Or weirder hacks. The fact that #set is also #initialize is bound
to make at least a few people cringe, but it works. In reality, I could
probably break that off into two functions wherein init calls set.
That's low-priority for now, though.