Remove unneeded class tests in bindings.
These were only really useful when using preprocessor-generated
functions since I wanted subclassing to work (rather than just returning
a new object of the base class type every time). Should still work. The
rule for this is fairly simple:

For binary ops where both lhs and rhs are the same, the output type is
lhs's type (the receiver's type). For binary ops where rhs and output
are the same, the output type is rhs's type unless output is provided,
in which case it's just writing to the output object.

For unary ops where the receiver's type is the same as the output type,
it's always the receiver's type. Otherwise, it's always the output base
class or simply written to output. Subclasses may need to override unary
ops if they want the output to be of a different type, possibly by doing
something like

  def unary_op(output = nil)
    super(output || NewOutputType.new)
  end

which is fundamentally the same thing that the bindings already do, only
it's moving the allocation elsewhere. Just make sure the base class of
the new output type is compatible with the original output type. So,
if output needs to be a Mat3, subclasses must inherit from Mat3 at some
point in their inheritance chain.