Add support for Marshal via _dump/_load.
Array support is currently hacked in since Marshal.dump appears to
traverse instance variables, which means dumping an element of an array
creates a circular reference because of how wrapped objects are set up
to prevent their parent memory from being garbage-collected while still
in use.

So, for now, it just duplicates the elements on _dump.