Add notes on thread-safety to the README.
In general, this is probably specific to my use case, but the point I'm
trying to get across is that none of this is designed with threads in
mind. That rhymes. Not intentional. So I just want to make it clear
what is and isn't safe, and it definitely isn't safe to modify the same
object from multiple threads at the same time. Or even really sane to
modify something from multiple threads.

Messaging is your friend. ZeroMQ included as a recommendation, though
not as something you must specifically use either.