Add some constants for identifying various properties of the gem.
Snow::SNOW_MATH_FLOAT_SIZE -> The size in bytes of the float type used
by snow-math. Either 4 or 8, depending on whether s_float_t aliases
float or double. With this, you should be able to either a) reject the
currently installed build of snow-math if it's not the right size float
or b) alter your code to accept differently-sized floats (e.g., switch
between GL_FLOAT and GL_DOUBLE). Personally, I recommend the first
option because trying to change things like this depending on the gem
environment isn't the best of ideas.

All math types now have SIZE and LENGTH constants that identify their
size in bytes and number of components. You could do this previously by
allocating one of the objects and getting the size and length from it,
but this is a little more convenient since it saves on an allocation.

All typed arrays now have a TYPE constant that reports their underlying
type. So, Mat3Array::TYPE is Mat3, Vec4Array::TYPE is Vec4, and so on.