Fix exception strings where method names were accidentally typenames.
This particular problem is mostly because, in the process of getting
rid of a lot of preprocessor stuff, I used snippets to build the new
code -- so, it was basically a preprocessor but in the editor. Upside:
made it really easy to get rid of the preprocessor. Downside: was still
more or less the same as using a preprocessor. So, things needed a lot
of tweaking.