Fix mat4_perspective output.
The perspective matrix was off by quite a bit in odd cases and I wasn't
quite sure why 'til I looked at this and compared it to gluPerspective's
matrix, at which point it became obvious that I must've been completely
out of it when I wrote the perspective function.

So yeah, that's fixed.