Fix array freezing.
Array elements weren't frozen, which broke freezing since you could
freeze the array but still modify its data by manipulating the elements.
Ordinarily, this would be fine, but the array data and the array
elements are the same memory, so it seems wrong to allow modifying them
by just modifying the element instead of using store to copy something
over the existing memory.