Make S_FLOAT_EPSILON a global variable rather than a constant.
This allows it to be more easily changed from within Ruby code, though
that's not done yet.

Also, for floats, use an epsilon of 1e-6, and for doubles, use 1e-9.
This fixes a check for whether a matrix is orthogonal, as 32-bit floats
would cause comparison of (m*m.tranpose == IDENTITY) to return false,
which is true if you're highly exact about it, but otherwise false if
if you're only concerned about so much precision (which is the case if
you're using floats in the first place).