Add pitch/yaw/roll methods to matrices.
This just returns the pitch, yaw, and roll of a matrix (in Mat4's case,
it only considers the inner 3x3 matrix), or at least what it might
describe if the rotations were reduced to Euler angles. This is
imprecise and I would generally caution anyone against using Euler
rotations when you've got tools like quaternions, but it's fairly
useful in some games.